requirements.txt
pyusb>=1.0.0  # USB communication with SpaceMouse
